--- a/sonic_platform_base/sonic_xcvr/utils/__init__.py
+++ b/sonic_platform_base/sonic_xcvr/utils/__init__.py
@@ -1,0 +1,1 @@
+# utils package for transceiver common utilities