… sonic_platform_base/sonic_xcvr/fields/consts.py
… sonic_platform_base/sonic_xcvr/mem_maps/public/cmis.py
… tests/sonic_xcvr/test_cmis_cache.py